feat: open product toggle
