+++ packages/api/src/window.d.ts
@@ -1,25 +1,1 @@
-declare global {
-  interface Window {
-    OIDC_URL: string;
-    OIDC_REALM: string;
-    OIDC_CLIENT_ID: string;
-    OIDC_REDIRECT_URI: string;
-    OIDC_POST_LOGOUT_REDIRECT_URI?: string;
-    GRAPHQL_URI: string;
-    REST_URI: string;
-    SHOW_INHABITANT_AMOUNT: string;
-    SHOW_ADDRESS_RESEARCH: string;
-    ADDRESS_RESEARCH_URL: string;
-    ACCOUNT_SHOW_NOTIFICATION_SUBSECTION: string;
-    OVERVIEW_SHOW_INTRO: string;
-    OVERVIEW_SHOW_MAINTENANCE_ALERT: string;
-    OVERVIEW_SHOW_CURRENT_CASES_PREVIEW: string;
-    OVERVIEW_CURRENT_CASES_PREVIEW_LENGTH: string;
-    THEME_CLASS: string;
-    MESSAGE_COUNT_POLLING_INTERVAL: number;
-    CASES_PARTIAL_SEARCH: boolean;
-    OPEN_KLANT_VERSION: string;
-  }
-}
-
-export {};
+import "../../../window";
