+++ packages/app/public/config.template.js
@@ -11,6 +11,7 @@ window.MESSAGE_COUNT_POLLING_INTERVAL = Number(
 );
 window.CASES_PARTIAL_SEARCH = "${CASES_PARTIAL_SEARCH}" === "true";
 window.OPEN_KLANT_VERSION = "${OPEN_KLANT_VERSION}";
+window.OPEN_PRODUCTEN = "${OPEN_PRODUCTEN}" === "true";
 window.ADDRESS_RESEARCH_URL = "${ADDRESS_RESEARCH_URL}";
 window.REPORT_CHANGE_OF_ADDRESS_URL = "${REPORT_CHANGE_OF_ADDRESS_URL}";
 window.CHANGE_IN_USE_OF_SURNAME_URL = "${CHANGE_IN_USE_OF_SURNAME_URL}";
