+++ packages/app/src/constants/config.ts
@@ -28,4 +28,5 @@ export const config = {
   MESSAGE_COUNT_POLLING_INTERVAL: window.MESSAGE_COUNT_POLLING_INTERVAL,
   CASES_PARTIAL_SEARCH: window.CASES_PARTIAL_SEARCH,
   OPEN_KLANT_VERSION: window.OPEN_KLANT_VERSION,
+  OPEN_PRODUCTEN: window.OPEN_PRODUCTEN,
 };
